Move HistorySummary to own file and class
